Move league read and management routes through facades and fill the league serializer
Extracts the league listing, detail, results, settings, organizer
management, delete, manage, and invitations routes from _api.py into a
dedicated leagues blueprint.

Replaces the league_serializer facade with the real require_league and
league_to_dict implementations. _api.py and content.py now import those
public names from the serializer instead of the underscore-prefixed
originals, and the remaining _api.py callsites are renamed to match.

URL surface is unchanged.
